refactor: Dialog/toast + GameDetailClient 拆分 + 型別統一 + Medium 修復 (#69)
High:
- confirm() → AlertDialog（5 處）、alert() → toast()（17 處）
- GameDetailClient 860→140 行：拆成 5 個 Tab 元件 + useGameDetail hook + types/game.ts
- usePlanManager API 改為 pendingDelete state 模式

Medium:
- getExcerpt 統一到 utils.ts
- LEAGUE_OPTIONS 統一到 constants.ts
- QuickOdds URL 改用 route helper
- useRewritePolling interval cleanup
- PickCenter 進度條互補計算防溢出
- fetchWinProbability 移除死參數 sport
- PersonaForm/ChannelForm 加 form submit Enter 支援
- smoke-test.config.json 補 2 個 API
